make collapsible element reusable
