+++ app/scripts/modules/core/src/presentation/CollapsibleElement.less
@@ -1,0 +1,41 @@
+.collapsible-element {
+  position: relative;
+
+  .content {
+    position: relative;
+    overflow: hidden;
+
+    &.collapsed:after {
+      content: '';
+      width: 100%;
+      height: 100%;
+      position: absolute;
+      left: 0;
+      bottom: 0;
+      background: linear-gradient(0deg, var(--color-white) 0px, transparent 40px);
+    }
+  }
+  .expand-button {
+    display: none;
+    position: absolute;
+    bottom: 0;
+    text-align: center;
+    padding-top: 2px;
+    border: none;
+    border-top: 2px solid #009cb8;
+    color: #009cb8;
+    width: 100%;
+    background-color: var(--color-white);
+    outline: none;
+
+    &:hover {
+      text-decoration: underline;
+    }
+  }
+
+  &:hover {
+    .expand-button {
+      display: block;
+    }
+  }
+}
