+++ app/scripts/modules/core/src/presentation/index.ts
@@ -1,6 +1,7 @@
+export * from './CollapsibleElement';
 export * from './collapsibleSection/CollapsibleSection';
-export * from './details/Details';
 export * from './CustomLabels';
+export * from './details/Details';
 export * from './FormElements';
 export * from './forms';
 export * from './hooks';
@@ -20,8 +21,8 @@ export * from './RenderOutputFile';
 export * from './robotToHumanFilter/robotToHuman.filter';
 export * from './sortToggle';
 export * from './SpanDropdownTrigger';
-export * from './SpinErrorBoundary';
 export * from './spel';
+export * from './SpinErrorBoundary';
 export * from './TabBoundary';
 export * from './tables';
 export * from './TetheredCreatable';
